+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [ZeroVer](https://0ver.org/) versioning.
 - Make censors editable with ctrl + right click
 - Support for checkboxes in the settings
 - "Interact with censors out of censor mode" setting => if checked, you can delete/edit censor rects event when out of censor mode
+### Fixed
+- Dark theme for pfp and banner
 
 ## [0.6.0] - 2025-12-18
 ### Changed